pyproject.toml
[tool.poetry]
name = "remesh_clone"
version = "0.1.0"
description = "clone of remesh platform"
authors = ["John Wilkinson <johnpwilkinson@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
